Updated configs. Added BASIL-specific config. Added SCHNAIL staging.

+++ pw.py
@@ -16,7 +16,6 @@ def main():
   args = parser.parse_args()
   if args.all or args.jre:
     makejre()
-    copyjre()
   if args.all or args.stage:
     stage()
   if args.all or args.package:
@@ -28,15 +27,18 @@ def pathjoin(*args):
     result = result.replace('//', '/')
   return result  
 
-file_launch4j = "c:/Program Files (x86)/Launch4j/Launch4jc.exe"
-dir_jre       = "c:/p/graalvm-jdk"
-dir_bwapidata = "c:/p/bw/bwapi-data/"
-java_version  = "21"
-dir_pw        = "c:/p/pw"
-dir_bots      = "c:/Users/d/AppData/Roaming/scbw/bots/"
-dir_out       = pathjoin(dir_pw,  "out/")
-dir_configs   = pathjoin(dir_pw,  "configs/")
-dir_staging   = pathjoin(dir_out, "staging/")
+file_launch4j     = "c:/Program Files (x86)/Launch4j/Launch4jc.exe"
+dir_jre           = "c:/p/graalvm-jdk/"
+dir_bwapidata     = "c:/p/bw/bwapi-data/"
+java_version      = "21"
+dir_pw            = "c:/p/pw/"
+dir_scbwbots      = "c:/Users/d/AppData/Roaming/scbw/bots/"
+dir_localschnail  = "c:/Program Files (x86)/SCHNAIL Client/bots/PurpleWave/"
+dir_out           = pathjoin(dir_pw,  "out/")
+dir_configs       = pathjoin(dir_pw,  "configs/")
+dir_staging       = pathjoin(dir_out, "staging/")
+dir_jar           = pathjoin(dir_out, "artifacts", "PurpleWave")
+file_jar          = pathjoin(dir_jar, "PurpleWave.jar")
 
 def path_pw(path):
   return pathjoin(dir_pw, path)
@@ -46,12 +48,9 @@ def path_configs(path):
   return pathjoin(dir_configs, path)  
 def path_staging(path):
   return pathjoin(dir_staging, path)
-def path_bots(path):
-  return pathjoin(dir_bots, path)
+def path_scbwbots(path):
+  return pathjoin(dir_scbwbots, path)
   
-dir_jar   = pathjoin(dir_out, "artifacts", "PurpleWave")
-file_jar  = pathjoin(dir_jar, "PurpleWave.jar")
-
 def log(*args, **kwargs):
   print(*args, **kwargs, flush=True)
 def logf(f, *args, **kwargs):
@@ -60,7 +59,7 @@ def logf(f, *args, **kwargs):
   
 def rmtree(dir):
   if os.path.exists(dir):
-    shutil.rmtree(dir)
+    logf(shutil.rmtree, dir)
 
 def makejre():
   log()
@@ -92,14 +91,6 @@ def makejre():
       "--output",
       path_staging("jre") ])
 
-def copyjre():   
-  log()
-  log("COPYING JRE")
-  for race in ["Protoss", "Terran", "Zerg", "Random"]:
-    dir_jre = path_bots(f"Purple{race}/AI/jre")    
-    logf(rmtree, dir_jre)
-    logf(shutil.copytree, path_staging("jre"), dir_jre)
-
 def stage():
   log()
   log("STAGING")
@@ -120,27 +111,40 @@ def stage():
   logf(shutil.copy2, path_configs("PurpleWaveLocalMetal.config.json"), pathjoin(dir_bwapidata, "AI") )
   
   log()
-  log("Populate SC-Docker")  
-  for race in ["Protoss", "Terran", "Zerg", "Random"]:    
-    name = f"Purple{race}"
-    log(race)
-    log(name)
-    dir_bot_bwapidata = path_bots(name)
-    populate_bwapidata(dir_bot_bwapidata)
-    logf(shutil.copy2, path_configs("PurpleWaveLocalDocker.config.json"),  pathjoin(dir_bot_bwapidata, "ai"))    
-    # Customize bot.json
-    bot_json = pathjoin(dir_bot_bwapidata, "bot.json")
-    content = ""
-    with open(bot_json, 'r') as file:
-      content = file.read().replace("Protoss", race).replace("PurpleWave", name)
-    with open(bot_json, 'w') as file:
-      file.write(content)
+  if os.path.exists(dir_scbwbots):
+    log("Populate SC-Docker")  
+    for race in ["Protoss", "Terran", "Zerg", "Random"]:    
+      name = f"Purple{race}"
+      log(race)
+      log(name)
+      dir_bot_bwapidata = path_scbwbots(name)
+      populate_bwapidata(dir_bot_bwapidata)
+      logf(shutil.copy2, path_configs("bot.json"),                          dir_bot_bwapidata)
+      logf(shutil.copy2, path_configs("PurpleWaveLocalDocker.config.json"), pathjoin(dir_bot_bwapidata, "ai"))
+      # Customize bot.json
+      bot_json = pathjoin(dir_bot_bwapidata, "bot.json")
+      content = ""
+      with open(bot_json, 'r') as file:
+        content = file.read().replace("Protoss", race).replace("PurpleWave", name)
+      with open(bot_json, 'w') as file:
+        file.write(content)
+    else:
+      log("Did not find SC-Docker")
       
-def populate_bwapidata(dir_bwapidata):
-  dir_ai = pathjoin(dir_bwapidata, "ai")
+  log()
+  if (os.path.exists(dir_localschnail)):
+    log("Populate SCHNAIL")
+    populate_bwapidata(dir_localschnail, dir_localschnail)
+    logf(shutil.copy2, path_configs("PurpleWaveSCHNAIL.config.json"), dir_localschnail)
+  else:
+    log("Did not find SCHNAIL")
+      
+def populate_bwapidata(dir_bwapidata, dir_ai=None):
+  dir_ai = pathjoin(dir_bwapidata, "ai") if dir_ai is None else dir_ai
   log(dir_bwapidata)
   log(dir_ai)
-  logf(os.makedirs,                                   dir_ai, exist_ok=True)
+  logf(os.makedirs,                                   dir_ai, exist_ok=True)  
+  logf(shutil.copy2, path_configs("bwapi.dll"),       dir_bwapidata)
   logf(shutil.copy2, path_staging("PurpleWave.jar"),  dir_ai)
   logf(shutil.copy2, path_staging("PurpleWave.exe"),  dir_ai)
   logf(shutil.copy2, path_staging("timestamp.txt"),   dir_ai)
@@ -148,36 +152,29 @@ def populate_bwapidata(dir_bwapidata):
   logf(shutil.copy2, path_configs("run_proxy.bat"),   dir_ai)
   logf(shutil.copy2, path_configs("run64.bat"),       dir_ai)
   logf(shutil.copy2, path_configs("run32.bat"),       dir_ai)
-  logf(shutil.copy2, path_configs("bot.json"),        dir_bwapidata)
-  logf(shutil.copy2, path_configs("bwapi.dll"),       dir_bwapidata)
 
 def package():
   log()
   log("PACKAGING")
-  for package in ["AIIDE", "SSCAIT", "SCHNAIL"]:
+  for package in ["AIIDE", "BASIL", "SSCAIT", "SCHNAIL"]:
     package_name = f"PurpleWave{package}"
     package_dir  = path_staging(package_name)
     log(package_name)
     log(package_dir)
     logf(rmtree,        package_dir)
     logf(os.makedirs,   package_dir)
-    logf(shutil.copy2,  path_staging("PurpleWave.jar"),               package_dir)
-    logf(shutil.copy2,  path_staging("PurpleWave.exe"),               package_dir)
-    logf(shutil.copy2,  path_staging("timestamp.txt"),                package_dir)
-    logf(shutil.copy2,  path_staging("revision.txt"),                 package_dir)
+    populate_bwapidata(package_dir, package_dir)
     logf(shutil.copy2,  path_configs(f"{package_name}.config.json"),  package_dir)
-  
-  log()
-  log("Copying package-specific resources")
-  logf(shutil.copy2, path_configs("BWAPI.dll"),     path_staging("PurpleWaveSSCAIT"))
-  logf(shutil.copy2, path_configs("BWAPI.dll"),     path_staging("PurpleWaveSCHNAIL"))
-  logf(shutil.copy2, path_configs("run_proxy.bat"), path_staging("PurpleWaveSSCAIT"))
-  logf(shutil.copy2, path_configs("run_proxy.bat"), path_staging("PurpleWaveAIIDE"))
-  open(path_staging(pathjoin("PurpleWaveAIIDE", "PurpleWave.dll")), 'w').close()
+    if package in ["AIIDE", "SCHNAIL"]:
+      dir_package_jre = pathjoin(package_dir, "jre")    
+      logf(rmtree, dir_package_jre)
+      logf(shutil.copytree, path_staging("jre"), dir_package_jre)
+    if package in ["AIIDE"]:
+      open(pathjoin(package_dir, "PurpleWave.dll"), 'w').close()  
   
   log()
   log("Replacing zips")
-  for package in ["AIIDE", "SSCAIT", "SCHNAIL"]:
+  for package in ["AIIDE", "BASIL", "SSCAIT", "SCHNAIL"]:
     package_name = f"PurpleWave{package}"
     package_dir  = path_staging(package_name)
     log(package_name)
